Fix excessive CPU use on desktop
While this is not meant for use on desktop, when running top, I realized it was using 100% of the CPU for a fucking flash card application.

My initial belief was that it was related to the draw cycle.  Using Callgrind, I determined it was caused by something involving libnvidia-glcore/libglx_mesa.so.  After running the output through an agent, it determined that there was a busy-wait inside the GLX swap-buffer path.

I'm using X11/GLX so this wait happens in user space, while on the device (that is using wayland), it has a kernel-space blocking wait setup.

I added some code to demo_main that executes only on x86_64 and adds a wait if we are on X86_64.  In reality we probably want to use more specific logic but this is enough for my purposes (the desktop is only to validate it runs).

This did not affect the test device I have (A 35XX Plus).